Lock NPM version in server
The package-lock.json is built according to the old version 1
formatting. Newer versions of npm build using the version 2, which
results in big confusing diffs in the lockfile.

This commit locks the npm version by requiring a specific node version
(the current LTS version). This should work since most people use
whatever npm came with their node version rather than installing a
specific version separately.

The `engines` section in package.json _prevents_ `npm install` being run
with a different version, while the `.nvmrc` _allows_ users to switch to
that version when using this project.

Note that `.nvmrc` is recognised by other version management tools (eg
asdf), not just by nvm.

An alternative approach would be to specify a minimum node version and
to rebuild the package-lock.json file with the new formatting, but maybe
there's some reason not to do this? (eg. if the node version installed
on the server is older?)