Use recommended eslint settings and fix issues
This is the next step up in eslint usage.

Symlink the eslintignore file to the project's .gitignore
This is so that we don't run eslint over the server/dist directory.
There doesn't seem to be a good approach for this, but symlinking seems
to be what a number of people online recommend as a simple solution.

Delete spec/helpers/reporters.js
This had lint errors due to using "require". It also didn't seem to
work? My guess was that it was added at some point while experimenting
with test config, never worked, and was committed by accident?

Add type definitions to src/gcal.ts
The default recommended typescript rules don't like you to use explicit
any. The fact that the type here isn't a full WcjEvent (and therefore
not compatible with the current calendar?) makes me wonder if this is
just dead code which we should delete?

It's always there in git if we need it.